Refactor SubscriberTextField to TextField
These fields are now going to be reused by other non-subscriber filters

MAILPOET-4990
